migrate utils to dedicated submodule
